Rename paramiko.py
This prevents an issue the imports where it thinks we are wanting to import the file vs the module
