+++ mysocketctl/ssh/__init__.py
@@ -1,2 +1,7 @@
+from mysocketctl.ssh.paramiko_client import Paramiko
 from mysocketctl.ssh.system import SystemSSH
-from mysocketctl.ssh.paramiko import Paramiko
+
+__all__ = [
+    SystemSSH,
+    Paramiko,
+]
